test: move jest tests fixtures to a separate folder
